pyproject.toml
[tool.poetry]
name = "timberk_gendiff"
version = "0.3.0"
description = ""
authors = ["Lysak Tatiana <lysak-tatiana2501@ya.ru>"]
packages = [
    { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.7"
flake8 = "^3.7"
pytest = "^5.3"
pytest-cov = "^2.8"
PyYAML = "^5.2"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
